Added hero-image to individual photo project
This image will be used as a banner on the individual project page. For each project, the hero-image will be different and specified in the banner field of the JSON object.
- Fix styles for layout
Margins and paddings have been updated for better visual consistency.
- Fix styles for mobile devices
Styles have been optimized for mobile devices to improve readability and the visual experience on small screens. Issues with margins have been fixed, and icons in the navbar dropdown are now hidden.